+++ packages/hybrid-login-view/src/__fixtures__/hybridLoginView.js
@@ -1,0 +1,41 @@
+const hybridLoginView = {
+  howItWorksProps: {
+    header:
+      'One Big Switch takes the stress out of getting value on your household bills by doing the neogtiating for you!',
+    stepOffers: [
+      {
+        imgUrl: 'https://placehold.it/100x100',
+        title: 'You join the movement for free',
+      },
+      {
+        imgUrl: 'https://placehold.it/100x100',
+        title: 'We negotiate Group Discounts',
+      },
+      {
+        imgUrl: 'https://placehold.it/100x100',
+        title: 'You decide what’s right for you',
+      },
+    ],
+  },
+  disclaimerProps: {
+    disclaimerText: '* disclaimer text',
+  },
+  title: 'Login panel title',
+  authenticityToken:
+    'I7t3w13syZv9gWaz0kbQA3fL6NqrhwiZlfIt9HixtbKNZIy5nsB36XE8eQHC+AtA5lT2p7Kt182TIrUDgSZ+gw==',
+  hiddenFields: {
+    registering_campaign_id: 276,
+    jump_path: '/campaigns/OBS_BES_2017_Offers',
+  },
+  loginUrl: '/guest_sessions',
+  autocompletePostcodeUrl: '/suburbs/autocomplete_postcode',
+  footerProps: {
+    logoUrl: 'https://placehold.it/100x100',
+  },
+  headerProps: {
+    logoUrl: 'https://placehold.it/100x100',
+    heroImageUrl: 'https://placehold.it/1080x454',
+  },
+};
+
+export default hybridLoginView;