+++ packages/login-panel/src/LoginForm/index.js
@@ -2,13 +2,13 @@ import React from 'react';
 import t from 'prop-types';
 import { Formik, Field } from 'formik';
 import styled from 'styled-components';
-// import { getColor } from '@rtm-ui/theme';
+import { getColor } from '@rtm-ui/theme';
 import { Box } from '@rtm-ui/layout';
 import Button from '@rtm-ui/button';
 import { Header, Paragraph } from '@rtm-ui/typography';
+import Icon from '@rtm-ui/icon';
 
 import { submitLogin } from './actions';
-import { LOGIN_URL } from './constants';
 import PostCodeField from '../PostCodeField';
 import GdprAgreement from '../GdprAgreement';
 
@@ -20,13 +20,13 @@ const StyledInput = styled.input`
   width: 100%;
 `;
 
-const ErrorWrapper = styled(Paragraph)`
-  background: #faeded;
-  position: relative;
-  padding: 10px 20px;
-  margin: 5px auto 5px;
-  border: 1px solid #ca3838;
-  color: #7b2121;
+const Error = styled(Paragraph)`
+  text-align: center;
+  color: ${props => getColor('error', props.theme)};
+  svg {
+    fill: ${props => getColor('error', props.theme)};
+    padding-top: 2px;
+  }
 `;
 
 const ButtonWrapper = styled(Box)`
@@ -50,24 +50,33 @@ class LoginForm extends React.Component {
   }
 
   async handleSubmit(values, actions) {
-    const result = await submitLogin(
-      LOGIN_URL,
-      values,
-      this.props.authenticityToken
-    );
+    let result;
+    if (typeof this.props.handleSubmit === 'function') {
+      result = await this.props.handleSubmit(
+        this.props.loginUrl,
+        values,
+        this.props.authenticityToken
+      );
+    } else {
+      result = await submitLogin(
+        this.props.loginUrl,
+        values,
+        this.props.authenticityToken
+      );
+    }
 
-    if (result.errors) {
+    const { data } = result;
+
+    if (data.errors) {
       this.setState({
-        errors: result.errors,
+        errors: data.errors,
       });
       actions.setSubmitting(false);
     } else if (typeof this.props.handleSuccess === 'function') {
       this.props.handleSuccess(result);
-      actions.setSubmitting(true);
     } else {
       // Redirect to path when success login
-      window.location.href = result.redirectPath;
-      actions.setSubmitting(true);
+      window.location.href = data.redirectPath;
     }
   }
 
@@ -78,6 +87,7 @@ class LoginForm extends React.Component {
       title,
       buttonText,
       gdprProps,
+      autocompletePostcodeUrl,
     } = this.props;
 
     const { errors } = this.state;
@@ -110,13 +120,6 @@ class LoginForm extends React.Component {
               <Header pt={[2, 2, 3, 4]} tag="h6">
                 {title}
               </Header>
-              {errors && (
-                <Box py={2}>
-                  {errors.map(error => (
-                    <ErrorWrapper key={error}>{error}</ErrorWrapper>
-                  ))}
-                </Box>
-              )}
               <Box pb={2} pt={[2, 2, 3, 4]}>
                 <Paragraph py={2}>My Postcode:</Paragraph>
                 <Field
@@ -127,6 +130,7 @@ class LoginForm extends React.Component {
                       field={field}
                       inputComponent={TextInput}
                       authenticityToken={authenticityToken}
+                      autocompletePostcodeUrl={autocompletePostcodeUrl}
                       id="user.postcode_suburb"
                       aria-labelledby="user.postcode_suburb"
                       placeholder="Postcode"
@@ -157,6 +161,15 @@ class LoginForm extends React.Component {
                 <Button type="submit" disabled={isSubmitting} track="signin">
                   {buttonText}
                 </Button>
+                {errors && (
+                  <Box pt={2}>
+                    {errors.map(error => (
+                      <Error key={error}>
+                        <Icon glyph="error" size={20} /> {error}
+                      </Error>
+                    ))}
+                  </Box>
+                )}
               </ButtonWrapper>
               {<GdprAgreement {...gdprProps} />}
             </form>
@@ -168,8 +181,10 @@ class LoginForm extends React.Component {
 }
 
 LoginForm.propTypes = {
-  authenticityToken: t.string,
+  authenticityToken: t.string.isRequired,
+  loginUrl: t.string.isRequired,
   handleSuccess: t.func,
+  handleSubmit: t.func,
   // eslint-disable-next-line react/forbid-prop-types
   hiddenFields: t.object,
   title: t.string,
@@ -177,12 +192,13 @@ LoginForm.propTypes = {
   buttonIcon: t.string,
   // eslint-disable-next-line react/forbid-prop-types
   gdprProps: t.shape({}),
+  autocompletePostcodeUrl: t.string,
 };
 
 LoginForm.defaultProps = {
   title:
     'Join One Big Switch today for FREE and instantly unlock your special offers!',
-  buttonText: 'See the offer',
+  buttonText: 'See the offers',
   buttonIcon: null,
   showGdprAgreement: true,
 };