+++ packages/login-panel/src/__tests__/LoginPanel.js
@@ -1,55 +1,35 @@
 import React from 'react';
-import axios from 'axios';
 // eslint-disable-next-line import/named
 import {
   render,
   // eslint-disable-next-line import/named
-  fireEvent,
-  // eslint-disable-next-line import/named
-  wait,
-  // eslint-disable-next-line import/named
   cleanup,
 } from '../../../bootstrap/setup/testSetup';
 import LoginPanel from '../index';
+import loginPanelProps from '../__fixtures__/loginPanel';
 
 jest.mock('axios');
 
 // automatically unmount and cleanup DOM after the test is finished.
 afterEach(cleanup);
 
-const mockDefaultAxios = () => {
-  axios.get.mockResolvedValueOnce({ data: [] });
-  axios.post.mockResolvedValueOnce({ data: { redirectPath: '/' } });
-};
-
 describe('<LoginPanel />', () => {
   it('matches expected output', async () => {
-    // set Up
-    mockDefaultAxios();
-
-    const { getByText, getByPlaceholderText, container } = render(
-      <LoginPanel />
+    const { getByText, getByValue, container } = render(
+      <LoginPanel {...loginPanelProps} />
     );
 
-    const email = getByPlaceholderText('Email');
-    fireEvent.change(email, {
-      target: { value: 'user@example.com' },
-    });
-    const postcode = getByPlaceholderText('Postcode');
-    fireEvent.change(postcode, {
-      target: { value: '2000, Barangaroo' },
-    });
-
-    const submit = getByText('See the offer');
-    fireEvent.click(submit);
-
-    expect(container).toHaveTextContent(
-      'Join One Big Switch today for FREE and instantly unlock your special offers!'
+    // expect hidden fields
+    const jumpPath = getByValue(loginPanelProps.hiddenFields.jump_path);
+    expect(jumpPath.name).toEqual('jump_path');
+    const registeringCampaignId = getByValue(
+      loginPanelProps.hiddenFields.registering_campaign_id.toString()
     );
+    expect(registeringCampaignId.name).toEqual('registering_campaign_id');
 
-    // expect event was fired
-    await wait(() => {
-      expect(submit).toBeDisabled();
-    });
+    expect(getByText(loginPanelProps.title)).toBeInTheDocument();
+    expect(getByText(loginPanelProps.buttonText)).toBeInTheDocument();
+
+    expect(container).toMatchSnapshot();
   });
 });