+++ packages/login-panel/src/__tests__/PostCodeField.js
@@ -10,6 +10,7 @@ import {
   cleanup,
 } from '../../../bootstrap/setup/testSetup';
 import PostCodeField from '../PostCodeField';
+import loginPanelProps from '../__fixtures__/loginPanel';
 
 jest.mock('axios');
 
@@ -37,6 +38,7 @@ describe('<PostCodeField />', () => {
           options={[{ value: '5000, ADELAIDE', label: '5000, ADELAIDE' }]}
           form={form}
           field={field}
+          autocompletePostcodeUrl={loginPanelProps.autocompletePostcodeUrl}
           getAutoCompletePostcode={getAutoCompletePostcode}
         />
       </React.Fragment>