+++ packages/button/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.2](https://github.com/OBSAU/rtm-kit/compare/@rtm-ui/button@3.0.1...@rtm-ui/button@3.0.2) (2019-01-15)
+
+**Note:** Version bump only for package @rtm-ui/button
+
+
+
+
+
+## [3.0.1](https://github.com/OBSAU/rtm-kit/compare/@rtm-ui/button@1.0.9...@rtm-ui/button@3.0.1) (2019-01-15)
+
+**Note:** Version bump only for package @rtm-ui/button
+
+
+
+
+
+## [1.0.9](https://github.com/OBSAU/rtm-kit/compare/@rtm-ui/button@1.0.8...@rtm-ui/button@1.0.9) (2019-01-14)
+
+**Note:** Version bump only for package @rtm-ui/button
+
+
+
+
+
+## [1.0.8](https://github.com/OBSAU/rtm-kit/compare/@rtm-ui/button@1.0.7...@rtm-ui/button@1.0.8) (2019-01-10)
+
+**Note:** Version bump only for package @rtm-ui/button
+
+
+
+
+
+## [1.0.7](https://github.com/OBSAU/rtm-kit/compare/@rtm-ui/button@1.0.6...@rtm-ui/button@1.0.7) (2019-01-10)
+
+
+### Bug Fixes
+
+* name modules.js to module.js on build ([8a9848c](https://github.com/OBSAU/rtm-kit/commit/8a9848c))
+
+
+
+
+
 ## [1.0.6](https://github.com/OBSAU/rtm-kit/compare/@rtm-ui/button@1.0.5...@rtm-ui/button@1.0.6) (2018-12-19)
 
 **Note:** Version bump only for package @rtm-ui/button